corrigindo tipagem do modelo das tabelas

+++ model/City.py
@@ -2,9 +2,9 @@ from configuration.config import ormDatabase as orm
 from model.State import State
 
 class City(orm.Model):
-    id = orm.Column(orm.Integer, primary_key=True)
+    id = orm.Column(orm.BigInteger, primary_key=True)
     name = orm.Column(orm.String(200))
-    state_id = orm.Column(orm.Integer, orm.ForeignKey('state.id'))
+    state_id = orm.Column(orm.BigInteger, orm.ForeignKey('state.id'))
     ibge_id = orm.Column(orm.Integer)
     
     def __init__(self, name, state_id):
