+++ model/District.py
@@ -1,9 +1,9 @@
 from configuration.config import ormDatabase as orm
 
 class District(orm.Model):
-    id = orm.Column(orm.Integer, primary_key=True)
+    id = orm.Column(orm.BigInteger, primary_key=True)
     name = orm.Column(orm.String(200))
-    city_id = orm.Column(orm.Integer, orm.ForeignKey('city.id'))
+    city_id = orm.Column(orm.BigInteger, orm.ForeignKey('city.id'))
     def __init__(self, name, cityId):
         self.name = name
         self.city_id = cityId
